Doc saliency functions and move them into own folder
