+++ perturbationdrive/uncertainty/__init__.py
@@ -1,0 +1,7 @@
+from perturbationdrive.uncertainty.saliencymap import (
+    getSaliencyMap,
+    getSaliencyPixels,
+    getSaliencyRegions,
+    plotImageAndSaliencyMap,
+    plotSaliencyRegions,
+)
